some refactoring and tidying up
