+++ handlers/new-product-api/src/main/scala/com/gu/newproduct/api/addsubscription/Formatters.scala
@@ -1,9 +1,7 @@
 package com.gu.newproduct.api.addsubscription
 
-import com.gu.newproduct.api.productcatalog.AmountMinorUnits
-
 object Formatters {
   implicit class AmountOps(amountMinorUnits: AmountMinorUnits) {
-    def prettyPrint: String = (amountMinorUnits.value / BigDecimal(100)).bigDecimal.setScale(2).toPlainString
+    def formatted: String = (amountMinorUnits.value / BigDecimal(100)).bigDecimal.setScale(2).toPlainString
   }
 }
