+++ handlers/new-product-api/src/main/scala/com/gu/newproduct/api/addsubscription/email/SendConfirmationEmail.scala
@@ -2,17 +2,15 @@ package com.gu.newproduct.api.addsubscription.email
 
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
-
 import com.gu.i18n.Currency
 import com.gu.newproduct.api.addsubscription.zuora.GetContacts.BilltoContact
 import com.gu.newproduct.api.addsubscription.zuora.GetPaymentMethod.{DirectDebit, PaymentMethod}
-import com.gu.newproduct.api.addsubscription.ZuoraAccountId
-import com.gu.newproduct.api.productcatalog.AmountMinorUnits
+import com.gu.newproduct.api.addsubscription.{AmountMinorUnits, ZuoraAccountId}
 import com.gu.util.Logging
 import com.gu.util.apigateway.ApiGatewayResponse
 import com.gu.util.reader.AsyncTypes._
 import com.gu.util.reader.Types.ApiGatewayOp.{ContinueProcessing, ReturnWithResponse}
-
+import com.gu.newproduct.api.addsubscription.Formatters._
 import scala.concurrent.Future
 
 object SendConfirmationEmail extends Logging {
@@ -51,8 +49,6 @@ object SendConfirmationEmail extends Logging {
 
   def hyphenate(s: String) = s"${s.substring(0, 2)}-${s.substring(2, 4)}-${s.substring(4, 6)}"
 
-  def formatAmount(amount: AmountMinorUnits) = (amount.value / BigDecimal(100)).bigDecimal.stripTrailingZeros.toPlainString
-
   val firstPaymentDateFormat = DateTimeFormatter.ofPattern("EEEE, d MMMM yyyy")
 
   def toContributionFields(currentDate: LocalDate, data: ContributionsEmailData): Option[ContributionFields] = {
@@ -65,7 +61,7 @@ object SendConfirmationEmail extends Logging {
       ContributionFields(
         EmailAddress = email.value,
         created = currentDate.toString,
-        amount = formatAmount(data.amountMinorUnits),
+        amount = data.amountMinorUnits.formatted,
         currency = data.currency.glyph,
         edition = data.billTo.country.map(_.alpha2).getOrElse(""),
         name = data.billTo.firstName.value,
