+++ handlers/new-product-api/src/test/scala/com/gu/newproduct/api/addsubscription/AddSubscriptionRequestTest.scala
@@ -2,7 +2,7 @@ package com.gu.newproduct.api.addsubscription
 
 import java.time.LocalDate
 
-import com.gu.newproduct.api.productcatalog.{AmountMinorUnits, PlanId}
+import com.gu.newproduct.api.productcatalog.PlanId
 import com.gu.newproduct.api.productcatalog.PlanId.MonthlyContribution
 import org.scalatest.{FlatSpec, Matchers}
 import play.api.libs.json.{JsError, Json}
