+++ handlers/new-product-api/src/test/scala/com/gu/newproduct/api/addsubscription/validation/contribution/ContributionValidationsTest.scala
@@ -4,9 +4,9 @@ import java.time.LocalDate
 
 import com.gu.i18n.Currency
 import com.gu.i18n.Currency._
+import com.gu.newproduct.api.addsubscription.AmountMinorUnits
 import com.gu.newproduct.api.addsubscription.validation.contribution.ContributionValidations.ValidatableFields
 import com.gu.newproduct.api.addsubscription.validation.{AmountLimits, Failed, Passed, ValidationResult}
-import com.gu.newproduct.api.productcatalog.AmountMinorUnits
 import org.scalatest.{FlatSpec, Matchers}
 
 class ContributionValidationsTest extends FlatSpec with Matchers {
